Add CultureInfo support to DateTimeArgument parsing
The `DateTimeArgument` class now includes a `CultureInfo?` property to specify a culture for parsing date and time strings. The `TrySetValue` method has been updated to use this property, defaulting to `CultureInfo.CurrentCulture` if not set. Multiple `DateTime.TryParseExact` and `DateTime.TryParse` calls now use the specified or default culture instead of `CultureInfo.InvariantCulture`. The `DateTimeArgumentFixture` test class has been updated to create and assign a new `CultureInfo` instance for "en-US" and to set the `CultureInfo` property in the `TrySetValueAndVerifyValue` method.